+++ Caro Board Game/Bot.hpp
@@ -1,7 +1,11 @@
-﻿#ifndef BOT_HPP
+﻿// BOT_HPP
+#ifndef BOT_HPP
 #define BOT_HPP
 
 #include "Player.hpp"
+#include "Referee.hpp"
+#include "Board.hpp"
+#include "BotAlgorithm.hpp"
 #include <iostream>
 #include <vector>
 
@@ -11,18 +15,27 @@ class Bot : public Player {
 private:
     // --- Private Members ---
     vector<pair<int, int>> moves;
-
+    BotAlgorithm AI;
 public:
     // --- Constructor ---
     Bot(char symbol) : Player(symbol) {}
 
-    // --- Game Logic Functions ---
-    void addMove(int x, int y) {
-        // Máy tính chọn nước đi thông qua thuật toán Minimax hoặc AI
-    }
-
-    void resetMoves() {
-        if (!moves.empty()) moves.clear();
+    bool decideMove(sf::RenderWindow& window, sf::Event& event, Board& board, Referee& referee, int& isWin) override {
+        pair<int, int> dec = AI.findMove(board, referee, getSymbol());
+        int x = dec.first, y = dec.second;
+        board.setCell(x, y, getSymbol());
+        addMove(x, y);
+        cout << x << " " << y << "\n";
+        for (int i = 0; i < board.getN(); i++) {
+            for (int j = 0; j < board.getM(); j++) {
+                if (board.getCell(i, j) == ' ') cout << ".";
+                else cout << board.getCell(i, j);
+            }
+            cout << "\n";
+        }
+        referee.checkWin(board);
+        cout << "win: " << isWin << "\n";
+        return true;
     }
 };